data transformation pipeline working
