+++ data-prep/README.md
@@ -1,39 +1,21 @@
-## Fixing gleamviz data
-First fix `md_*.tsv` files which sometime countain two records on a single line. The script
-also removes unnecessary `ID` column to unify the format with the rest of the exported files.
-Then follow with a simple `sed` command removes leading space in all TSV files:
-
-```
-$ python remove_doubles.py <gleamviz-exported-data> data_fixed
-# removes # and leading space
-$ sed -i 's/#//;s/^.//' data_fixed/*/*.tsv
-```
-
-then you should be able to standardly load the data in pandas. Tested on linux.
-
-```python
-# in python shell
-import pandas as pd
-filename = "data_fixed/md_countries.tsv"  # or any other converted file
-df = pd.read_csv(filename, sep="\t", index_col=0)
-```
-
 ## Converting data into a single file
-On the result above, you can trigger
-```
-python convert_to_dfs.py data_fixed
-```
-which creates a Parquet file `area.pq` with all TSVs from different areas
-in a single dataframe with somewhat resonable memory usage (currently <200 MB).
 
-This can be loaded as:
+The data transformation pipeline assumes there is a folder with data from multiple GleamViz exported data
+
+In the root of the project run:
+```
+python data-prep/process_data.py folder_with_all_runs_data selected_cities_file [output_file]
+```
+which creates a hdf file with a default name `covid_data.hdf` containing selected preprocessed data + metadata.
+
+These can be loaded as:
 ```python
 import pandas as pd
-df = pd.read_parquet("area.pq")
-
-# and optionally set index to follow the directory structure via
-# df = df.set_index(["area_type", "source_main", "source_sec"])
+city_lookup = pd.read_hdf("covid_data.hdf", "data/city_lookup")
+metadata = pd.read_hdf("covid_data.hdf", "metadata/model_info")
 ```
+
+The `city_lookup` dataframe contains preprocessed timeseries from the simulation for each selected city. `metadata` currently contains only the date when the simulation was run (the first of all of the runs).
 
 ## Manual fixes
 ### `md_cities.tsv` issues
